Only setup TAESD preview on initial install if the config doesn't exist

+++ StabilityMatrix.Core/Models/Packages/A3WebUI.cs
@@ -179,10 +179,15 @@ public class A3WebUI : BaseGitPackage
         progress?.Report(new ProgressReport(1f, "Installing Package Requirements", isIndeterminate: false));
         
         progress?.Report(new ProgressReport(-1f, "Updating configuration", isIndeterminate: true));
+        
         // Create and add {"show_progress_type": "TAESD"} to config.json
+        // Only add if the file doesn't exist
         var configPath = Path.Combine(InstallLocation, "config.json");
-        var config = new JsonObject {{"show_progress_type", "TAESD"}};
-        await File.WriteAllTextAsync(configPath, config.ToString()).ConfigureAwait(false);
+        if (!File.Exists(configPath))
+        {
+            var config = new JsonObject {{"show_progress_type", "TAESD"}};
+            await File.WriteAllTextAsync(configPath, config.ToString()).ConfigureAwait(false);
+        }
 
         progress?.Report(new ProgressReport(1f, "Install complete", isIndeterminate: false));
     }
